% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{dark}
\alias{dark}
\title{Enable dark mode of a ggplot2 theme}
\usage{
dark(.theme, background = "black", geom_fill = "white",
  geom_color = "white", geom_colour = geom_color)
}
\value{
theme
}
\description{
Enable dark mode of a ggplot2 theme
}
