% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode_geoms.R
\name{dark_mode_geoms}
\alias{dark_mode_geoms}
\alias{get_geoms}
\alias{invert_geom_defaults}
\alias{geoms_are_ggdefault}
\alias{darken_geoms}
\alias{lighten_geoms}
\title{Manipulate geoms for dark mode}
\usage{
get_geoms()

invert_geom_defaults(geoms = get_geoms(), verbose = FALSE)

geoms_are_ggdefault(geoms = get_geoms())

darken_geoms(geoms = get_geoms(), verbose = FALSE)

lighten_geoms(geoms = get_geoms(), verbose = FALSE)
}
\arguments{
\item{geoms}{List of geoms as ggproto objects.}

\item{verbose}{Whether to print a message whenever inversion occurs.}
}
\description{
These functions help change default colors for color and fill aesthetics for
all geoms in the loaded namespaces that have them.
}
\details{
\code{get_geoms} Returns all the geoms from loaded namespaces.

\code{invert_geom_defaults} inverts geom defaults for fill and color.

\code{geoms_are_ggdefault} checks if the geoms are at their ggplot default
values for color and fill.

\code{darken_geoms} sets defaults to be dark-mode friendly.

\code{lighten_geoms} sets defaults to the ggplot2 defaults.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~Species)

p + dark_theme_gray() # geom defaults changed

p + theme_gray() # oh no! geoms are not visible on light background

lighten_geoms() # geom defaults changed back

p + theme_gray() # back to normal

}
