% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode_theme.R
\name{dark_mode_theme}
\alias{dark_mode_theme}
\alias{invert_theme_elements}
\alias{theme_is_dark}
\alias{darken_theme}
\alias{lighten_theme}
\title{Manipulate themes for dark mode}
\usage{
invert_theme_elements(.theme)

theme_is_dark(.theme)

darken_theme(.theme)

lighten_theme(.theme)
}
\arguments{
\item{.theme}{A ggplot theme}
}
\description{
\code{invert_theme_elements} returns the inverted theme
(blank elements will be made black).
\code{theme_is_dark} checks if the theme is dark.
\code{darken_theme} darkens theme (unless it's dark already).
\code{lighten_theme} lightens theme (unless it's light already).
}
