% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\docType{package}
\name{ggdark}
\alias{ggdark}
\alias{ggdark-package}
\title{ggdark: Dark mode for ggplot2 themes}
\description{
The only function you're likely to need from ggdark is [darken()]
which activates dark mode on your favorite ggplot2 theme. See the
README at github.com/nsgrantham/ggdark for more information.
}
