% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_dark.R
\name{ggsave_dark}
\alias{ggsave_dark}
\title{Save a dark/light version of ggplot}
\usage{
ggsave_dark(filename, plot = last_plot(), dark = FALSE, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{dark}{Whether to save as dark (\code{TRUE}) or light (\code{FALSE}).}

\item{...}{Other parameters passed to \code{\link[ggplot2]{ggsave}}}
}
\description{
Save a dark/light version of ggplot
}
