% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{greyscale_complement_hex}
\alias{greyscale_complement_hex}
\title{Get greyscale complement to color by hexcode}
\usage{
greyscale_complement_hex(color)
}
\arguments{
\item{color}{Greyscale color to complement (hexcode)}
}
\value{
Complementary greyscale color (hexcode)
}
\description{
Get greyscale complement to color by hexcode
}
\examples{
# Wrapped by greyscale_compelement
greyscale_complement("#333333")  # calls greyscale_complement_hex

}
