% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{greyscale_complement_name}
\alias{greyscale_complement_name}
\title{Get greyscale complement to color by name}
\usage{
greyscale_complement_name(color)
}
\arguments{
\item{color}{Greyscale color to complement (name)}
}
\value{
Complementary greyscale color (name)
}
\description{
Get greyscale complement to color by name
}
\examples{
# Wrapped by greyscale_compelement
greyscale_complement("grey80")  # calls greyscale_complement_name

}
