% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{invert_color}
\alias{invert_color}
\alias{invert_colour}
\title{Invert color(s)}
\usage{
invert_color(color, colour = color)

invert_colour(color, colour = color)
}
\arguments{
\item{color}{color(s) to invert}

\item{colour}{alias of color}
}
\value{
Inverted color(s) in hex code
}
\description{
Invert a vector of colors, provided the colors are valid hex codes
or have valid names (i.e., they belong to base::colors()), and
return a vector of inverted colors in hex code.
}
\examples{
invert_color("white")    # "black"
invert_color("gray20")   # "gray80"
invert_color("grey80")   # "grey20"
invert_color(c("#000000", "#333333"))  # "#FFFFFF","#CCCCCC"

}
