% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{invert_geom_defaults}
\alias{invert_geom_defaults}
\title{Inverse geom defaults for fill and color/colour}
\usage{
invert_geom_defaults(geoms = get_geoms())
}
\arguments{
\item{geoms}{List of geoms as ggproto objects}
}
\description{
Inverse geom defaults for fill and color/colour
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~ Species)

old <- theme_set(dark_mode())  # color and fill changed to "white"
p

update_geom_colors(color = "grey80", fill = "grey80")  # or perhaps a light grey
p

theme_set(old)
p  # oh no! geom defaults are still "white"

update_geom_colors()  # color and fill changed to "black"
p  # back to normal

}
