% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{invert_theme_elements}
\alias{invert_theme_elements}
\title{Invert theme elements}
\usage{
invert_theme_elements(.theme)
}
\arguments{
\item{.theme}{theme to invert}
}
\value{
Inverted theme
}
\description{
Invert theme elements
}
