% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{update_geom_colors}
\alias{update_geom_colors}
\title{Update geom defaults for fill and color/colour}
\usage{
update_geom_colors(fill, color, colour = color)
}
\arguments{
\item{fill}{Default fill value for ggplot2::geom_*}

\item{color}{Default color value for ggplot2::geom_*}

\item{colour}{Alias of color}
}
\description{
Update geom defaults for fill and color/colour
}
