% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{dark_mode}
\alias{dark_mode}
\title{Activate dark mode on a ggplot2 theme}
\usage{
dark_mode(theme = theme_get(), geom_colour = "white",
  geom_color = geom_colour, geom_fill = "white")
}
\arguments{
\item{theme}{ggplot2 theme object}

\item{geom_colour}{alias of geom_color}

\item{geom_color}{new geom color}

\item{geom_fill}{new geom fill}
}
\value{
dark version of theme
}
\description{
Activate dark mode on a ggplot2 theme
}
\examples{
library(ggplot2)

p1 <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point(aes(color = Species))

p1  # theme_gray(), the default
p1 + dark_mode()  # activate dark mode on theme_gray()

p2 <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~ Species)

p2 + dark_mode()  # geom color is white
p2 + dark_mode(geom_color = "grey80")  # or perhaps a light grey

update_geom_colors()  # restore defaults fill = "black", color = "black"

}
