% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{darken}
\alias{darken}
\title{Enable dark mode of a ggplot2 theme}
\usage{
darken(.theme = theme_get(), background = "black",
  geom_fill = "white", geom_color = "white",
  geom_colour = geom_color)
}
\arguments{
\item{.theme}{ggplot2 theme object}

\item{background}{Background fill}

\item{geom_fill}{Default fill value for ggplot2::geom_*}

\item{geom_color}{Default color value for ggplot2::geom_*}

\item{geom_colour}{Alias of geom_color}
}
\value{
darkened_theme
}
\description{
Enable dark mode of a ggplot2 theme
}
