% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{darken}
\alias{darken}
\title{Enable dark mode of a ggplot2 theme}
\usage{
darken(.theme = ggplot2::theme_get(), background = "black",
  geom_fill = "white", geom_color = "white",
  geom_colour = geom_color, update_geoms = TRUE)
}
\arguments{
\item{.theme}{ggplot2 theme object}

\item{background}{Background fill}

\item{geom_fill}{Default fill value for ggplot2::geom_*}

\item{geom_color}{Default color value for ggplot2::geom_*}

\item{geom_colour}{Alias of geom_color}

\item{update_geoms}{Overwrite default fill and color values for ggplot2::geom_*?}
}
\value{
Theme with dark mode elements
}
\description{
Enable dark mode of a ggplot2 theme
}
\examples{
library(ggplot2)

p1 <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
geom_point(aes(color = Species))

p1  # theme_gray(), the default
p1 + darken()  # activate dark mode on theme_gray()
p1 + darken(theme_minimal())  # darken a different ggplot2 theme
p1 + darken(theme_minimal(), background = "grey20")  # lighten black background

p2 <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~ Species)

p2 + darken()  # geom color is now white
p2 + darken(geom_color = "grey80")  # or perhaps a light grey

# you can update the geoms outside of darken, but make
# sure they're not overwritten when darken is run again
update_geom_colors(color = "orange", fill = "gold")
p2 + darken(update_geoms = FALSE)

restore_geom_colors()  # restore defaults fill = "black", color = "black"

}
