% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{greyscale_complement}
\alias{greyscale_complement}
\title{Get complement on grey scale}
\usage{
greyscale_complement(color)
}
\arguments{
\item{color}{Greyscale color for which to find its complement}
}
\description{
Get complement on grey scale
}
