% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{greyscale_complement}
\alias{greyscale_complement}
\title{Get greyscale complement}
\usage{
greyscale_complement(color)
}
\arguments{
\item{color}{Greyscale color to complement}
}
\description{
Get greyscale complement
}
