% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{greyscale_complement_hex}
\alias{greyscale_complement_hex}
\title{Get greyscale complement to color by hexcode}
\usage{
greyscale_complement_hex(color)
}
\arguments{
\item{color}{Greyscale color to complement}
}
\description{
Get greyscale complement to color by hexcode
}
