% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{geoms_are_ggdefault}
\alias{geoms_are_ggdefault}
\alias{invert_geom_defaults}
\alias{set_geom_defaults}
\title{Change geom defaults}
\usage{
geoms_are_ggdefault(geoms = get_geoms())

invert_geom_defaults(geoms = get_geoms())

set_geom_defaults(dark, geoms = get_geoms())
}
\arguments{
\item{geoms}{List of geoms as ggproto objects}

\item{dark}{Whether to reset geoms to be dark mode friendly. If \code{TRUE},
makes defaults white (to show on black background, the ggplot default). If
\code{FALSE}, makes them black (to show on default white background).}
}
\description{
\code{invert_geom_defaults} inverts geom defaults for fill and color.
}
\details{
\code{set_geom_defaults} sets geom defaults to the ggplot default or to be
dark-mode friedly.

\code{geoms_are_ggdefault} checks if the geoms are at their ggplot default
values for color and fill.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~Species)

p + dark_theme_gray() # geom defaults changed

p + theme_gray() # oh no! geoms are not visible on light background

set_geom_defaults(dark = FALSE) # geom defaults changed back

p + theme_gray() # back to normal
}
