% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdark.R
\name{restore_geom_colors}
\alias{restore_geom_colors}
\alias{restore_geom_colours}
\title{Restore the geom defaults of fill = "white and color = "black"}
\usage{
restore_geom_colors()

restore_geom_colours()
}
\description{
Restore the geom defaults of fill = "white and color = "black"
}
