% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_mode.R
\name{update_geom_colors}
\alias{update_geom_colors}
\alias{update_geom_colours}
\title{Update geom defaults for fill and color/colour}
\usage{
update_geom_colors(color = "black", colour = color, fill = "black",
  verbose = TRUE)

update_geom_colours(color = "black", colour = color, fill = "black",
  verbose = TRUE)
}
\arguments{
\item{color}{new geom color}

\item{colour}{alias of color}

\item{fill}{new geom fill}

\item{verbose}{print messages (default: TRUE)}
}
\value{
List of geoms and their new defaults for fill and color/colour (invisibly returned)
}
\description{
Update geom defaults for fill and color/colour
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point() +
  facet_wrap(~ Species)

old <- theme_set(dark_mode())  # color and fill changed to "white"
p

update_geom_colors(color = "grey80", fill = "grey80")  # or perhaps a light grey
p

theme_set(old)
p  # oh no! geom defaults are still "white"

update_geom_colors()  # color and fill changed to "black"
p  # back to normal

}
