% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{update_geom_colors}
\alias{update_geom_colors}
\title{Update geom defaults for fill and color/colour}
\usage{
update_geom_colors(fill, color, colour = color)
}
\description{
Update geom defaults for fill and color/colour
}
